// src/api/tracking/events/TrackingEvent.ts
export enum TrackingEvent {
    // Standard events
    COMPLETE_REGISTRATION = 'COMPLETE_REGISTRATION',
    VIEW_CONTENT = 'VIEW_CONTENT',
    SEARCH = 'SEARCH',
    INITIATE_CHECKOUT = 'INITIATE_CHECKOUT',
    PURCHASE = 'PURCHASE',
    SUBSCRIBE = 'SUBSCRIBE',
    ADD_TO_CART = 'ADD_TO_CART',
    LEAD = 'LEAD',

    // Custom events
    RECIPE_GENERATED = 'RECIPE_GENERATED',
    RECIPE_GENERATION_INITIATED = 'RECIPE_GENERATION_INITIATED',
    RECIPE_GENERATION_FAILED = 'RECIPE_GENERATION_FAILED',
    QUOTA_LIMIT_REACHED = 'QUOTA_LIMIT_REACHED',
    CREDIT_PACK_VIEWED = 'CREDIT_PACK_VIEWED',
    FRIDGE_FINISHED_STEP3 = 'FRIDGE_FINISHED_STEP3',
    INSTAGRAM_IMPORT_STARTED = 'INSTAGRAM_IMPORT_STARTED',
    INSTAGRAM_IMPORT_FINISHED = 'INSTAGRAM_IMPORT_FINISHED',
}